Moves ArraySortHelper common code to shared location
I made the entry methods static as only CoreClr specific optimization needs them
as an instance methods.

This will require small CoreRT tweak

-                ArraySortHelper<TKey, TValue>.Default.Sort(keys, items, index, length, comparer);
+                ArraySortHelper<TKey, TValue>.Sort(keys, items, index, length, comparer);

which removes accessing following field which I don't really know much about

https://github.com/dotnet/corert/blob/57289ff33dba98c8aea472019adebc08ec68634b/src/System.Private.CoreLib/src/System/Collections/Generic/ArraySortHelper.cs#L328-L332
